refactor client pages to allow metadata
